problems/binary_search/E704_BinarySearch_test.go
package binary_search

import (
	"testing"

	"github.com/ruijzhan/leetCode/pkg/utils"
)

func Test_search(t *testing.T) {
	fs := []func([]int, int) int{search, search2, search3}
	type args struct {
		nums   []int
		target int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			args: args{
				nums:   []int{-1, 0, 3, 5, 9, 12},
				target: 9,
			},
			want: 4,
		},
		{
			args: args{
				nums:   []int{-1, 0, 3, 5, 9, 12},
				target: 2,
			},
			want: -1,
		},
		{
			args: args{
				nums:   []int{-1, 0, 5},
				target: -1,
			},
			want: 0,
		},
		{
			args: args{
				nums:   []int{-1, 0, 3, 5, 9, 12},
				target: 13,
			},
			want: -1,
		},
	}
	for _, tt := range tests {
		for _, f := range fs {
			t.Run(tt.name, func(t *testing.T) {
				if got := f(tt.args.nums, tt.args.target); got != tt.want {
					t.Errorf("%v() = %v, want %v", utils.GetFunctionName(f), got, tt.want)
				}
			})
		}
	}
}
